Assignment5

+++ app/src/main/java/com/example/gl_crash_course/api/model/Main.kt
@@ -7,7 +7,7 @@ data class Main(
     @SerializedName("humidity")
     val humidity: Int,
     @SerializedName("pressure")
-    val pressure: Int,
+    val pressure: Double,
     @SerializedName("temp")
     val temp: Double,
     @SerializedName("temp_max")
